renamed code projects from Pricing to Prices; directory still Pricing
