+++ docker-compose.yml
@@ -1,7 +1,7 @@
 services:
 
     esdb:
-        image: eventstore/eventstore:24.2.0-bookworm-jammy
+        image: eventstore/eventstore:24.2.0-jammy
         container_name: ecomm_esdb
         ports:
             - '2113:2113'
@@ -98,14 +98,14 @@ services:
 #            context: .
 #            dockerfile: src/Inventory/Inventory.Api/Dockerfile
 
-#    pricing.api:
-#        image: pricing.api
-#        container_name: ecomm_pricing_api
+#    prices.api:
+#        image: prices.api
+#        container_name: ecomm_prices_api
 #        ports:
 #            - "5219:80"
 #        build:
 #            context: .
-#            dockerfile: src/Pricing/Pricing.Api/Dockerfile
+#            dockerfile: src/Pricing/Prices.Api/Dockerfile
 
 #    discounts.api:
 #        image: discounts.api
