+++ src/Pricing/Prices.Api/Dockerfile
@@ -1,0 +1,23 @@
+﻿FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
+USER $APP_UID
+WORKDIR /app
+EXPOSE 8080
+EXPOSE 8081
+
+FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
+ARG BUILD_CONFIGURATION=Release
+WORKDIR /src
+COPY ["src/Pricing/Prices.Api/Prices.Api.csproj", "src/Pricing/Prices.Api/"]
+RUN dotnet restore "src/Pricing/Prices.Api/Prices.Api.csproj"
+COPY . .
+WORKDIR "/src/src/Pricing/Prices.Api"
+RUN dotnet build "Prices.Api.csproj" -c $BUILD_CONFIGURATION -o /app/build
+
+FROM build AS publish
+ARG BUILD_CONFIGURATION=Release
+RUN dotnet publish "Prices.Api.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "Prices.Api.dll"]
