+++ src/Pricing/Prices.Api/appsettings.Development.json
