+++ src/Pricing/Prices.Api/appsettings.json
